+++ htmlresources/alxmlportwizard/js/alxmlportwizard.js
@@ -1,7 +1,7 @@
 class XmlPortWizard extends TableBasedObjectWizard {
 
     constructor() {
-        super(2);
+        super(2, false);
 
         this.registerFieldsSelectionEvents();